% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dic.R
\name{get_dic}
\alias{get_dic}
\title{Get Dictionary}
\usage{
get_dic(dic_name)
}
\arguments{
\item{dic_name}{one of dictionary name(character), \strong{woorimalsam}, \strong{insighter}, \strong{sejong}}
}
\value{
The \code{data.frame} object contains tags and terms
}
\description{
Get Dictionary
}
\examples{
\dontrun{
   dic_df <-get_dic('sejong')
}
}

